backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
assemblyai==0.26.0
av==12.0.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
ctranslate2==4.2.1
decorator==4.4.2
deep-translator==1.11.4
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.3
faster-whisper==1.0.2
ffmpeg-python==0.2.0
filelock==3.14.0
flatbuffers==24.3.25
fsspec==2024.3.1
future==1.0.0
google==3.0.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-speech==2.26.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imageio==2.34.1
imageio-ffmpeg==0.4.9
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.4.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.17.3
orjson==3.10.3
packaging==24.0
pillow==10.3.0
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.18.0
pymongo==4.7.2
pysrt==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.10
requests==2.31.0
rfc3986==1.5.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
sentencepiece==0.2.0
shellingham==1.5.4
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
SpeechRecognition==3.10.4
srt==3.5.3
starlette==0.37.2
sympy==1.12
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
tqdm==4.66.4
transformers==4.40.2
triton==2.3.0
typer==0.12.3
typing_extensions==4.11.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vosk==0.3.45
watchfiles==0.21.0
websockets==12.0
